Unfinished store; Implement WindowList
