+++ modules/windows/index.js
@@ -2,5 +2,12 @@ module.exports = {
     TokenWindow: require('./TokenWindow'),
     SettingsWindow: require('./SettingsWindow'),
     AppWindow: require('./AppWindow'),
-    SplashWindow: require('./SplashWindow')
+    SplashWindow: require('./SplashWindow'),
+
+    WindowList: {
+        AppWindow: null,
+        SplashWindow: null,
+        TokenWindow: null,
+        SettingsWindow: null
+    }
 }
